+++ SDL/Quintuplet.cu
@@ -185,30 +185,30 @@ void SDL::createQuintupletsInUnifiedMemory(struct SDL::quintuplets& quintupletsI
     quintupletsInGPU.tripletIndices = (unsigned int*)cms::cuda::allocate_managed(nMemoryLocations * 2 * sizeof(unsigned int), stream);
     quintupletsInGPU.lowerModuleIndices = (unsigned int*)cms::cuda::allocate_managed(nMemoryLocations * 5 * sizeof(unsigned int), stream);
     quintupletsInGPU.nQuintuplets = (unsigned int*)cms::cuda::allocate_managed(nLowerModules * sizeof(unsigned int), stream);
-    quintupletsInGPU.innerRadius = (float*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.outerRadius = (float*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.pt = (float*)cms::cuda::allocate_managed(nMemoryLocations *7* sizeof(float), stream);
+    quintupletsInGPU.innerRadius = (__half*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.outerRadius = (__half*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.pt = (__half*)cms::cuda::allocate_managed(nMemoryLocations *7* sizeof(__half), stream);
     quintupletsInGPU.layer = (int*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(int), stream);
     quintupletsInGPU.isDup = (bool*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(bool), stream);
     quintupletsInGPU.partOfPT5 = (bool*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(bool), stream);
-    quintupletsInGPU.regressionRadius = (float*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.regressionG = (float*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.regressionF = (float*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(float), stream);
+    quintupletsInGPU.regressionRadius = (__half*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.regressionG = (__half*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.regressionF = (__half*)cms::cuda::allocate_managed(nMemoryLocations * sizeof(__half), stream);
 
 #else
     cudaMallocManaged(&quintupletsInGPU.tripletIndices, 2 * nMemoryLocations * sizeof(unsigned int));
     cudaMallocManaged(&quintupletsInGPU.lowerModuleIndices, 5 * nMemoryLocations * sizeof(unsigned int));
 
     cudaMallocManaged(&quintupletsInGPU.nQuintuplets, nLowerModules * sizeof(unsigned int));
-    cudaMallocManaged(&quintupletsInGPU.innerRadius, nMemoryLocations * sizeof(float));
-    cudaMallocManaged(&quintupletsInGPU.outerRadius, nMemoryLocations * sizeof(float));
-    cudaMallocManaged(&quintupletsInGPU.regressionRadius, nMemoryLocations * sizeof(float));
-    cudaMallocManaged(&quintupletsInGPU.pt, nMemoryLocations *4* sizeof(float));
+    cudaMallocManaged(&quintupletsInGPU.innerRadius, nMemoryLocations * sizeof(__half));
+    cudaMallocManaged(&quintupletsInGPU.outerRadius, nMemoryLocations * sizeof(__half));
+    cudaMallocManaged(&quintupletsInGPU.regressionRadius, nMemoryLocations * sizeof(__half));
+    cudaMallocManaged(&quintupletsInGPU.pt, nMemoryLocations *4* sizeof(__half));
     cudaMallocManaged(&quintupletsInGPU.layer, nMemoryLocations * sizeof(int));
     cudaMallocManaged(&quintupletsInGPU.isDup, nMemoryLocations * sizeof(bool));
     cudaMallocManaged(&quintupletsInGPU.partOfPT5, nMemoryLocations * sizeof(bool));
-    cudaMallocManaged(&quintupletsInGPU.regressionG, nMemoryLocations * sizeof(float));
-    cudaMallocManaged(&quintupletsInGPU.regressionF, nMemoryLocations * sizeof(float));
+    cudaMallocManaged(&quintupletsInGPU.regressionG, nMemoryLocations * sizeof(__half));
+    cudaMallocManaged(&quintupletsInGPU.regressionF, nMemoryLocations * sizeof(__half));
 
 #ifdef CUT_VALUE_DEBUG
     cudaMallocManaged(&quintupletsInGPU.innerRadiusMin, nMemoryLocations * sizeof(float));
@@ -253,15 +253,15 @@ void SDL::createQuintupletsInExplicitMemory(struct SDL::quintuplets& quintuplets
     quintupletsInGPU.tripletIndices = (unsigned int*)cms::cuda::allocate_device(dev, 2 * nMemoryLocations * sizeof(unsigned int), stream);
     quintupletsInGPU.lowerModuleIndices = (unsigned int*)cms::cuda::allocate_device(dev, 5 * nMemoryLocations * sizeof(unsigned int), stream);
     quintupletsInGPU.nQuintuplets = (unsigned int*)cms::cuda::allocate_device(dev, nLowerModules * sizeof(unsigned int), stream);
-    quintupletsInGPU.innerRadius = (float*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.outerRadius = (float*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.pt = (float*)cms::cuda::allocate_device(dev, nMemoryLocations *7* sizeof(float), stream);
+    quintupletsInGPU.innerRadius = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.outerRadius = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.pt = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations *7* sizeof(__half), stream);
     quintupletsInGPU.layer = (int*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(int), stream);
     quintupletsInGPU.isDup = (bool*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(bool), stream);
     quintupletsInGPU.partOfPT5 = (bool*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(bool), stream);
-    quintupletsInGPU.regressionRadius = (float*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.regressionG = (float*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(float), stream);
-    quintupletsInGPU.regressionF = (float*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(float), stream);
+    quintupletsInGPU.regressionRadius = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.regressionG = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(__half), stream);
+    quintupletsInGPU.regressionF = (__half*)cms::cuda::allocate_device(dev, nMemoryLocations * sizeof(__half), stream);
 
 #else
     //cudaMallocAsync(&quintupletsInGPU.tripletIndices, 2 * nMemoryLocations * sizeof(unsigned int),stream);
@@ -279,15 +279,15 @@ void SDL::createQuintupletsInExplicitMemory(struct SDL::quintuplets& quintuplets
     cudaMalloc(&quintupletsInGPU.tripletIndices, 2 * nMemoryLocations * sizeof(unsigned int));
     cudaMalloc(&quintupletsInGPU.lowerModuleIndices, 5 * nMemoryLocations * sizeof(unsigned int));
     cudaMalloc(&quintupletsInGPU.nQuintuplets, nLowerModules * sizeof(unsigned int));
-    cudaMalloc(&quintupletsInGPU.innerRadius, nMemoryLocations * sizeof(float));
-    cudaMalloc(&quintupletsInGPU.outerRadius, nMemoryLocations * sizeof(float));
-    cudaMalloc(&quintupletsInGPU.regressionRadius, nMemoryLocations * sizeof(float));
-    cudaMalloc(&quintupletsInGPU.pt, nMemoryLocations *4* sizeof(float));
+    cudaMalloc(&quintupletsInGPU.innerRadius, nMemoryLocations * sizeof(__half));
+    cudaMalloc(&quintupletsInGPU.outerRadius, nMemoryLocations * sizeof(__half));
+    cudaMalloc(&quintupletsInGPU.regressionRadius, nMemoryLocations * sizeof(__half));
+    cudaMalloc(&quintupletsInGPU.pt, nMemoryLocations *4* sizeof(__half));
     cudaMalloc(&quintupletsInGPU.isDup, nMemoryLocations * sizeof(bool));
     cudaMalloc(&quintupletsInGPU.partOfPT5, nMemoryLocations * sizeof(bool));
     cudaMalloc(&quintupletsInGPU.layer, nMemoryLocations * sizeof(int));
-    cudaMalloc(&quintupletsInGPU.regressionG, nMemoryLocations * sizeof(float));
-    cudaMalloc(&quintupletsInGPU.regressionF, nMemoryLocations * sizeof(float));
+    cudaMalloc(&quintupletsInGPU.regressionG, nMemoryLocations * sizeof(__half));
+    cudaMalloc(&quintupletsInGPU.regressionF, nMemoryLocations * sizeof(__half));
 #endif
     cudaMemsetAsync(quintupletsInGPU.nQuintuplets,0,nLowerModules * sizeof(unsigned int),stream);
     cudaStreamSynchronize(stream);
@@ -317,20 +317,17 @@ __device__ void SDL::addQuintupletToMemory(struct SDL::quintuplets& quintupletsI
     quintupletsInGPU.lowerModuleIndices[5 * quintupletIndex + 2] = lowerModule3;
     quintupletsInGPU.lowerModuleIndices[5 * quintupletIndex + 3] = lowerModule4;
     quintupletsInGPU.lowerModuleIndices[5 * quintupletIndex + 4] = lowerModule5;
-    quintupletsInGPU.innerRadius[quintupletIndex] = innerRadius;
-    quintupletsInGPU.outerRadius[quintupletIndex] = outerRadius;
-    quintupletsInGPU.regressionRadius[quintupletIndex] = regressionRadius;
-    quintupletsInGPU.pt[quintupletIndex] = pt;
-    quintupletsInGPU.eta[quintupletIndex] = eta;
-    quintupletsInGPU.phi[quintupletIndex] = phi;
-    //quintupletsInGPU.score_rphi[quintupletIndex] = scores[0];
-    //quintupletsInGPU.score_rz[quintupletIndex] = scores[1];
-    //quintupletsInGPU.score_rzlsq[quintupletIndex] = scores[3];
-    quintupletsInGPU.score_rphisum[quintupletIndex] = scores;
+    quintupletsInGPU.innerRadius[quintupletIndex] = __float2half(innerRadius);
+    quintupletsInGPU.outerRadius[quintupletIndex] = __float2half(outerRadius);
+    quintupletsInGPU.regressionRadius[quintupletIndex] = __float2half(regressionRadius);
+    quintupletsInGPU.pt[quintupletIndex] = __float2half(pt);
+    quintupletsInGPU.eta[quintupletIndex] = __float2half(eta);
+    quintupletsInGPU.phi[quintupletIndex] = __float2half(phi);
+    quintupletsInGPU.score_rphisum[quintupletIndex] = __float2half(scores);
     quintupletsInGPU.layer[quintupletIndex] = layer;
     quintupletsInGPU.isDup[quintupletIndex] = false;
-    quintupletsInGPU.regressionG[quintupletIndex] = regressionG;
-    quintupletsInGPU.regressionF[quintupletIndex] = regressionF;
+    quintupletsInGPU.regressionG[quintupletIndex] = __float2half(regressionG);
+    quintupletsInGPU.regressionF[quintupletIndex] = __float2half(regressionF);
 
 #ifdef CUT_VALUE_DEBUG
     quintupletsInGPU.innerRadiusMin[quintupletIndex] = innerRadiusMin;
